arm64: DT: sdm845-camera: Enable dynamic IB bandwidth update for HF AXI
In case of a sudden reconfiguration of the camera NoC we may get
stuck with the old IB bandwidth, which may not be sufficient,
producing stuttering during video recording or preview.

To avoid this, enable dynamic IB bandwidth updating for the real-time
AXI ports, ensuring that the right bw gets set and also getting
some very small power consumption benefits.