FROMLIST: HID: Support Microsoft SPI touch digitizer driver
Microsoft uses a touch digitizer that communicates to the main SoC via SPI
and presents itself as a HID device. This patch contains the changes needed
for the driver to work as a module: HID Core affordances for SPI devices,
addition of the new Device IDs, and a new quirk in hid-microsoft.

Bug: 194673479
Link: https://lore.kernel.org/linux-input/20210812001250.1709418-1-dmanti@microsoft.com/
Signed-off-by: Dmitry Antipov <dmanti@microsoft.com>
Change-Id: I413022a7a2e8151c002f3fb92f9e902a29f0974c
Signed-off-by: Rudi Setiyawan <diphons@gmail.com>