Revert "genetlink: add CAP_NET_ADMIN test for multicast bind"
This reverts commit 30cc13fe89401588a52a0aa148dceff8f3796486.

It does not have a direct commit relation in upstream, and was lovingly
hand-crafted just for the 4.19.y kernel tree with the hope that it would
be useful to someone.

Unfortunately for us, it breaks the Android kernel abi and so, if it is
really needed in the future, must be brought back in an abi-safe way
then.

Signed-off-by: Rudi Setiyawan <diphons@gmail.com>